refactor : getUserClient -> getUserId로 이름 수정
