+++ src/entities/user/api/getUserById.ts
