+++ src/features/user/me/hooks/useMeForm.ts
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { getUserById } from '@/entities/user/api/getUserClient';
+import { getUserById } from '@/entities/user/api/getUserById';
 import { MeFormData } from '../model/type';
 import { meFormSchema } from '../model/schema';
 import { updateUser } from '@/entities/user/api/updateUser';
